+++ samplers/jaegerremote/internal/ratelimiter/rate_limiter_test.go
@@ -16,7 +16,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package utils
+package ratelimiter
 
 import (
 	"testing"
